Remove all float conversions from uint8 path
ScoredEdge.score, Merge.score, and threshold parameters now use
ScoreValue (uint8_t in uint8 mode) instead of double/float. The entire
uint8 pipeline stays in integer arithmetic — no float anywhere.

+++ src/waterz/agglomerate.pyx
@@ -105,12 +105,12 @@ cdef extern from "frontend_agglomerate.h":
         uint64_t a
         uint64_t b
         uint64_t c
-        double score
+        float score
 
     struct ScoredEdge:
         uint64_t u
         uint64_t v
-        double score
+        float score
 
     struct WaterzState:
         int     context
